extract quick check evaluation module
