+++ src/lib/quickCheck/evaluation/types.ts
@@ -1,0 +1,11 @@
+import type { ReviewQuestionRetrievalResult, ReviewQuestionResult } from "@/lib/quickCheck/retrieval/types";
+
+export type EvaluateRetrievedReviewQuestionInput = Pick<
+  ReviewQuestionRetrievalResult,
+  "reviewArea" | "matchedHeadings" | "headingIndex" | "rejectedMatches" | "matchStage"
+>;
+
+export type ReviewQuestionEvaluationResult = Pick<
+  ReviewQuestionResult,
+  "baselineReview" | "reviewAreaReview" | "status"
+>;
